fix(#3435): address review — port Strategy tests, fix typo + redirect
CI unit failure: the deleted Strategy service still had a test suite
(tests/Unit/.../Strategy/Services/StrategyServiceTest.php) importing the
removed class. Ported its coverage to BlueprintsServiceTest (board metadata,
buildRecentProgressCanvas, and getBoardsOverview via a partial mock of the
now self-called getLastUpdatedCanvas/getBoardProgress) and deleted the stale
Strategy test directory.

Also addressed two Copilot review comments:
- Fixed the 'lastlastUpdate' placeholder typo in the board metadata map ->
  'lastUpdate' (the key the rest of the code and template actually read).
- The /strategy/showBoards legacy redirect now preserves the query string,
  matching the other legacy redirects in routes.php.
